:lipstick: footer 디자인 변경 (#63)
