+++ src/components/layout/order/footer/StepButton.tsx
@@ -1,16 +1,56 @@
 import React from 'react';
+import { SwiperRef } from 'swiper/react';
+import useOrderStore, { ORDER_STEP } from '@/shared/store/order';
+import { usePaymentWidget } from '@/hooks/order/usePaymentWidget';
+import styles from './styles/StepButton.module.css';
+import { convertNumberToWon } from '@/shared/helper';
+import CartIcon from '@/components/icons/CartIcon';
 
 interface ButtonProps {
-  text: string;
-  handler: React.MouseEventHandler;
-  disabled?: boolean;
+  sliderRef: React.RefObject<SwiperRef>;
 }
 
-const StepButton = ({ text, handler, disabled = false }: ButtonProps) => {
+export const SLIDE_MOVE_SPEED = 600;
+
+const StepButton = ({ sliderRef }: ButtonProps) => {
+  const { goNextStep, goPrevStep, orderList, step, getTotalPrice } = useOrderStore();
+  const { paymentWidget, handlePaymentRequest } = usePaymentWidget();
+
+  const BUTTON_OPTIONS: { [key: number]: string } = {
+    1: convertNumberToWon(getTotalPrice()),
+    2: '결제 하러 이동',
+    3: '결제 하기',
+  };
+
+  const prevClickHandler = () => {
+    sliderRef.current!.swiper.slidePrev(SLIDE_MOVE_SPEED);
+    goPrevStep();
+  };
+
+  const nextClickHandler = async () => {
+    if (step === ORDER_STEP.PAYMENT && paymentWidget) {
+      await handlePaymentRequest(orderList);
+    } else {
+      sliderRef.current!.swiper.slideNext(SLIDE_MOVE_SPEED);
+      goNextStep();
+    }
+  };
+
   return (
-    <button onClick={handler} disabled={disabled}>
-      {text}
-    </button>
+    <>
+      {step !== 0 && (
+        <div className={styles.container}>
+          <button className={styles.button} onClick={nextClickHandler} disabled={orderList.length === 0}>
+            <span>{step > ORDER_STEP.CHOOSE_ORDER_TYPE && BUTTON_OPTIONS[step]}</span>
+            {step === ORDER_STEP.SELECT_MENU && (
+              <div className={styles.iconWrapper}>
+                <CartIcon amount={orderList.length} />
+              </div>
+            )}
+          </button>
+        </div>
+      )}
+    </>
   );
 };
