+++ src/components/layout/order/footer/styles/Footer.module.css
@@ -2,6 +2,9 @@
     position: fixed;
     width: 100%;
     height: var(--kiosk-footer-height);
+    max-width: 76.8rem  /* 768/10 */;
     bottom: 0;
     z-index: var(--level-1-z-index);
+    padding: 0 1.25rem;
+    margin-bottom: 15px;
 }
