--- src/components/layout/order/footer/styles/StepButtonContainer.module.css
@@ -1,22 +1,0 @@
-.container {
-    display: flex;
-    height: 100%;
-    justify-content: space-around;
-    align-items: center;
-}
-
-.container button {
-    border: 0.1rem  /* 1/10 */ solid black;
-    border-radius: 1rem  /* 10/10 */;
-    padding: 2rem  /* 20/10 */;
-    background: #864AF9;
-    color: white;
-    cursor: pointer;
-    width: 20rem  /* 200/10 */;
-    margin: 2rem  /* 20/10 */;
-}
-
-.container button:disabled {
-    background: #cccccc;
-    border: 0.1rem  /* 1/10 */ solid #cccccc;
-}
