+++ src/styles/globals.css
@@ -1,6 +1,6 @@
 :root {
   --kiosk-header-height: 6rem  /* 80/10 */;
-  --kiosk-footer-height: 8rem  /* 80/10 */;
+  --kiosk-footer-height: 50px;
   --kiosk-container-padding: 1.25rem;
 
   --main-color: #8f00ff;
